+++ README.md
@@ -1,0 +1,1 @@
+`npm run server`
