+++ src/modules/client-adm/facade/client-adm.facade.spec.ts
@@ -1,6 +1,6 @@
 import { Sequelize } from "sequelize-typescript";
 import ClientAdmFacadeFactory from "../factory/client-adm.facade.factory";
-import { ClientModel } from "../repository/client.model";
+import ClientModel from "../repository/client.model";
 import ClientRepository from "../repository/client.repository";
 import AddClientUseCase from "../usecase/add-client/add-client.usecase";
 import FindClientUseCase from "../usecase/find-client/find-client.usecase";
