+++ src/modules/client-adm/repository/client.repository.spec.ts
@@ -1,7 +1,7 @@
 import { Sequelize } from "sequelize-typescript";
 import Id from "../../@shared/domain/value-object/id.value-object";
 import Client from "../domain/client.entity";
-import { ClientModel } from "./client.model";
+import ClientModel from "./client.model";
 import ClientRepository from "./client.repository";
 
 describe("ClientRepository test", () => {
