+++ src/modules/client-adm/repository/client.repository.ts
@@ -2,7 +2,7 @@ import Id from "../../@shared/domain/value-object/id.value-object";
 import Client from "../domain/client.entity";
 import clientEntity from "../domain/client.entity";
 import ClientGateway from "../gateway/client.gateway";
-import { ClientModel } from "./client.model";
+import ClientModel from "./client.model";
 
 export default class ClientRepository implements ClientGateway {
   async add(client: clientEntity): Promise<void> {
